requirements.txt
numpy
tf-nightly-2.0-preview
matplotlib
opencv-python
scikit-learn==0.20.0
pyyaml